Fix prints_len

+++ compiler/ras/Logger.cpp
@@ -47,7 +47,7 @@ OMR::Logger::Logger()
 int32_t OMR::Logger::prints_len(const char *string)
 {
     int32_t rc = prints(string);
-    return (rc >= 0) ? strlen(string) : rc;
+    return (rc >= 0) ? static_cast<int32_t>(strlen(string)) : rc;
 }
 
 int32_t OMR::Logger::printc_len(char c)
